+++ tests/index.spec.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const { genepi, GenepiReader } = require('../')
+const { genepi } = require('../')
 
 const outputter = {
   header: function header() {},
@@ -28,58 +28,6 @@ describe('Genepi', () => {
     spyInner.mockRestore()
     spyFooter.mockReset()
     spyFooter.mockRestore()
-  })
-
-  describe('GenepiReader', () => {
-    it('uses an outputter class [integration]', () => {
-      const reader = new GenepiReader()
-      const text = 'This is a splendid textstring.'
-
-      return reader.play(text, outputter, 1).then(() => {
-        expect(spyHeader).toHaveBeenCalled()
-        expect(spyInner).toHaveBeenCalled()
-        expect(spyFooter).toHaveBeenCalled()
-        expect(spyInner.mock.calls).toEqual([
-          ['This', 1],
-          ['is', 1],
-          ['a', 0],
-          ['splendid', 2],
-          ['textstring.', 3]
-        ])
-      })
-    })
-    it('can play forward from a given word index', () => {
-      const reader = new GenepiReader()
-      const text = 'This is a splendid textstring.'
-
-      return reader.play(text, outputter, 1, 2).then(() => {
-        expect(spyHeader).toHaveBeenCalled()
-        expect(spyInner).toHaveBeenCalled()
-        expect(spyFooter).toHaveBeenCalled()
-        expect(spyInner.mock.calls).toEqual([
-          ['a', 0],
-          ['splendid', 2],
-          ['textstring.', 3]
-        ])
-      })
-    })
-    it('can play backward', () => {
-      const reader = new GenepiReader()
-      const text = 'This is a splendid textstring.'
-
-      return reader.play(text, outputter, 1, 3, -1).then(() => {
-        expect(spyHeader).toHaveBeenCalled()
-        expect(spyInner).toHaveBeenCalled()
-        expect(spyFooter).toHaveBeenCalled()
-        expect(spyInner).not.toBeCalledWith('textstring.', 3)
-        expect(spyInner.mock.calls).toEqual([
-          ['splendid', 2],
-          ['a', 0],
-          ['is', 1],
-          ['This', 1]
-        ])
-      })
-    })
   })
 
   describe('genepi function iterates on a string containing words', () => {
@@ -130,40 +78,6 @@ describe('Genepi', () => {
 
       jest.advanceTimersByTime(200)
       expect(spyFooter).toBeCalled()
-    })
-  })
-  describe('GenepiReader advanced', () => {
-    it('can be paused and resumed', () => {
-      jest.useFakeTimers()
-      const text = 'This is a splendid textstring.'
-
-      const reader = new GenepiReader()
-      reader.play(text, outputter)
-
-      expect(spyInner).not.toBeCalled()
-
-      jest.advanceTimersByTime(400)
-      expect(spyInner.mock.calls).toEqual([
-        ['This', 1],
-        ['is', 1]
-      ])
-      const currentIndex = reader.pause()
-      expect(currentIndex).toBe(2)
-      jest.advanceTimersByTime(1000)
-
-      expect(spyInner.mock.calls).toEqual([
-        ['This', 1],
-        ['is', 1]
-      ])
-      reader.resume(text, outputter)
-      jest.advanceTimersByTime(1000)
-      expect(spyInner.mock.calls).toEqual([
-        ['This', 1],
-        ['is', 1],
-        ['a', 0],
-        ['splendid', 2],
-        ['textstring.', 3]
-      ])
     })
   })
 })