+++ CHANGELOG.md
@@ -1,2 +1,7 @@
 # Changelog
 ## 0.1.0 POC
+## 1.0.0 First version
+## 1.1.0 GenepiReader class returning promise
+Provide a first version of the API
+## 2.0.0 GenepiReader class returning promise
+Change API (GenepiReader class) contract to get handy functions